Rename tiledb-serde to tiledb-pod
