+++ test-utils/cells/Cargo.toml
@@ -9,13 +9,13 @@ paste = { workspace = true }
 proptest = { workspace = true }
 strategy-ext = { workspace = true }
 tiledb-common = { workspace = true }
+tiledb-pod = { workspace = true }
 tiledb-proptest-config = { workspace = true, optional = true }
-tiledb-serde = { workspace = true }
 
 [dev-dependencies]
 tiledb-common = { workspace = true, features = ["proptest-strategies"] }
-tiledb-serde = { workspace = true, features = ["proptest-strategies"] }
+tiledb-pod = { workspace = true, features = ["proptest-strategies"] }
 
 [features]
 default = []
-proptest-strategies = ["dep:tiledb-proptest-config", "tiledb-common/proptest-strategies", "tiledb-serde/proptest-strategies"]
+proptest-strategies = ["dep:tiledb-proptest-config", "tiledb-common/proptest-strategies", "tiledb-pod/proptest-strategies"]
