+++ tiledb/api/src/tests/examples/sparse_all.rs
@@ -8,10 +8,10 @@ use tiledb_common::array::dimension::DimensionConstraints;
 use tiledb_common::array::{ArrayType, CellValNum};
 use tiledb_common::datatype::Datatype;
 use tiledb_common::physical_type_go;
-use tiledb_serde::array::attribute::AttributeData;
-use tiledb_serde::array::dimension::DimensionData;
-use tiledb_serde::array::domain::DomainData;
-use tiledb_serde::array::schema::SchemaData;
+use tiledb_pod::array::attribute::AttributeData;
+use tiledb_pod::array::dimension::DimensionData;
+use tiledb_pod::array::domain::DomainData;
+use tiledb_pod::array::schema::SchemaData;
 
 pub type FnAcceptDimension = dyn Fn(&Parameters, Datatype) -> bool;
 pub type FnAcceptAttribute =
